v0.1 - wrong answer
It couldn't be done by naive binary search. For instance, if n = 5, as
a naive binary search, we pick {3, 4} in order as the worst case,
however, we could actually pick 4 at first, and if higher, 5 is the
number. Otherwise it's between 1 to 3. Then we choose 2. So the total
expense is 4 + 2 = 6, not $7.