v1.0 - iteration
The swift version!
I begin to learn swift. I plan to learn this programming language to
implement a Poker game. At the time of learning, I think I can use it
to reimplement the problems on Leetcode once again. So I not only learn
a new language, but also have a project owned by myself! It’s thrilled!